+++ hnote/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 **Added**
 
--   Possibility of state changes from the list level.
+-   Capabilities of deleting tasks and state changes, from the list level.
 
 ## V. 0.2.3 - April 05, 2020